DMARCParser

Documentation for DMARCParser.

DMARCParser.dmarc

DMARCParser.dmarc_formatted

DMARCParser.get_dmarc_data

DMARCParser.get_policy_data

DMARCParser.get_record_data

DMARCParser.get_report_data

DMARCParser.initialize_dataframe

DMARCParser.initialize_dictionary

DMARCParser.iterate_node

DMARCParser.iterate_record_node

DMARCParser.populate_dictionary

DMARCParser.dmarc — Method

dmarc(xml_file_name_and_path)

This function processes a DMARC XML report and returns the data in a DataFrame.

Arguments

xml_file_name_and_path::String: The file name and path of the DMARC XML file.

Returns

A DataFrame containing the unformatted (i.e. raw) DMARC data.

DMARCParser.dmarc_formatted — Method

dmarc_formatted(xml_file_name_and_path)

Converts a DMARC report stored in XML format to a formatted DataFrame.

This function reads the DMARC report from the specified XML file, processes the data, and returns a formatted DataFrame with the following modifications:

Converts 'begin' and 'end' columns to DateTime objects.

Converts 'adkim' and 'aspf' columns from "r" to "Relaxed" and "s" to "Strict".

Adds a unique ID column combining row numbers and report IDs.

Arguments

xml_file_name_and_path::AbstractString: The file name and path of the DMARC XML file.

Returns

DataFrame: Formatted DataFrame with the processed DMARC file.

DMARCParser.get_dmarc_data — Method

get_dmarc_data(xml_file_name_and_path)

Parse a DMARC (Domain-based Message Authentication, Reporting, and Conformance) XML file and extract relevant information from its sections.

Parameters

xml_file_name_and_path::String: The path and filename to the DMARC XML file.

Returns

A dictionary containing consolidated DMARC data

Example

xml_file_name_and_path = "data/dmarc.xml"
dmarc_data = get_dmarc_data(xml_file_name_and_path)

DMARCParser.get_policy_data — Method

get_policy_data(policy_elements_vector, number_of_records)

Extracts and processes specified tag names from the Policy section of the DMARC XML file.

Arguments

policy_elements_vector::Vector{Node}: Vector of XML-like elements representing the Policy data.

number_of_records::Int64: Number of record sets in the XML data

Returns

A dictionary containing the specified policy tag names as keys and their corresponding values.

DMARCParser.get_record_data — Method

get_record_data(record_elements_vector, number_of_records)

Extracts and processes specified tag names from the Record sections of the DMARC XML file.

Arguments

record_elements_vector::Vector{Node}: Vector of XML-like elements representing the Policy data.

number_of_records::Int64: Number of record sets in the XML data

Returns

A dictionary containing the specified record tag names as keys and their corresponding values.

DMARCParser.get_report_data — Method

get_report_data(report_elements_vector, number_of_records)

Extracts and processes specified tag names from the Report section of the DMARC XML file.

Arguments

report_elements_vector::Vector{Node}: Vector of XML-like elements representing the Report data.

number_of_records::Int64: Number of record sets in the XML data

Returns

A dictionary containing the specified report tag names as keys and their corresponding values.

DMARCParser.initialize_dataframe — Method

initialize_dataframe(array_of_columns)

Create a new DataFrame with empty columns based on the provided array_of_columns.

Arguments

array_of_columns::Vector{String}: A vector of String representing the names of the columns to be created in the DataFrame.

Returns

DataFrame: A new DataFrame with empty columns matching the provided column names.

DMARCParser.initialize_dictionary — Function

initialize_dictionary(array_of_keys, number_of_records = 1)

Create and initialize a dictionary with keys from array_of_keys and values as arrays of missing values.

Arguments

array_of_keys::Vector{String}: An array of unique keys that will be used as dictionary keys.

number_of_records::Int = 1: (Optional) The number of records (size of value arrays) to initialize for each key. Default is 1.

Returns

blank_dictionary::Dict{String, Vector{Union{Missing, String}}}: A dictionary with keys from array_of_keys, and values initialized as arrays of Missing values. Each value array has a size of number_of_records.

Notes

The array of Missing values is important for indicating which entries are missing in the final output.

Example

keys = ["name", "age", "city"]
dict = initialize_dictionary(keys, 3)
# Output:
# Dict{String, Vector{Union{Missing, String}}} with 3 entries:
#   "name" => [missing, missing, missing]
#   "age"  => [missing, missing, missing]
#   "city" => [missing, missing, missing]

This function is useful for creating an empty dictionary with specified keys and initializing the associated arrays with missing values.

DMARCParser.iterate_node — Function

iterate_node(node::XML.Node, required_tag_names::Vector{String}, results_array::Vector{Tuple{String, String}}=[])

Iterates through an XML document tree represented by node and extracts tag-value pairs for elements with tag names found in required_tag_names. It recursively explores the tree structure to handle nested elements.

Arguments

node::XML.Node: The XML node representing the root of the document or subtree.

required_tag_names::Vector{String}: A vector of tag names to be extracted from the XML document.

results_array::Vector{Tuple{String, String}}=[]: (Optional) An array to store the extracted tag-value pairs. Default is an empty array.

Returns

This function does not return a value but populates the results_array with the extracted tag-value pairs.

Notes

This function is intended for processing the Report and Policy nodes only.

The condition length(children(child)) > 0 is used to handle tags that are empty, such as <sp></sp>.

DMARCParser.iterate_record_node — Function

iterate_record_node(node::XML.Node, required_tag_names::Vector{String}, results_array::Vector{Tuple{String, String}}=[], parent_tag::String="")

Recursively iterate through an XML document node and extract specified tag-value pairs.

This function traverses an XML document node and extracts tag-value pairs based on specified criteria. It recursively explores child nodes, and when it finds a matching tag, it extracts the corresponding value. The extracted tag-value pairs are stored in a results array.

Arguments

node::XML.Node: The XML document node to start the iteration from.

required_tag_names::Vector{String}: A vector of tag names to extract values for.

results_array::Vector{Tuple{String, String}}=[]: (Optional) An array to store the extracted tag-value pairs.

parent_tag::String="": (Optional) The parent tag name for the current node, used to create unique tag names for nested results.

Returns

This function does not return a value but populates the results_array with the extracted tag-value pairs.

Notes

This function is for extracting tag-value pairs from the Record XML nodes ondly.

The condition length(children(child)) > 0 is used to handle tags that are empty, such as <sp></sp>.

This function addresses the issue where the <result> tags may not be unique by modifying the tag names to be unique based on the parent tag.

DMARCParser.populate_dictionary — Method

populate_dictionary(source_results_array, destination_dictionary)

Populates a destination dictionary with values from a source results array. This function iterates over the source results array, where each record is an array of tuple key-value pairs. It extracts the keys and values from each record and assigns the values to the corresponding key in the destination dictionary, associating them with the corresponding index.

Arguments

source_results_array::Vector{Vector{Tuple{String, String}}}: The source array containing records in the form of an array of tuple key-value pairs.

destination_dictionary::Dict{String, Vector{Union{Missing, String}}}: The dictionary to populate with the values from the source array.

Returns

Nothing: The function modifies the destination_dictionary in place.

Notes

The function assumes that the destination_dictionary already exists and is properly initialized before calling this function.

Example

report_results_array = [[("org_name", "Outlook.com"), ("email", "dmarcreport@microsoft.com"), ("report_id", "c9a3a6172b14410d884df1ac115c670e"), ("begin", "1646697600"), ("end", "1646784000")], 
    [("org_name", "Outlook.com"), ("email", "dmarcreport@microsoft.com"), ("report_id", "c9a3a6172b14410d884df1ac115c670e"), ("begin", "1646697600"), ("end", "1646784000")]]
populate_dictionary(report_results_array, report_dictionary)
# Output:
# Dict{String, Vector{Union{Missing, String}}} with 5 entries:
#   "begin"     => ["1646697600", "1646697600"]
#   "email"     => ["dmarcreport@microsoft.com", "dmarcreport@microsoft.com"]
#   "org_name"  => ["Outlook.com", "Outlook.com"]
#   "end"       => ["1646784000", "1646784000"]
#   "report_id" => ["c9a3a6172b14410d884df1ac115c670e", "c9a3a6172b14410d884df1ac115c670e"]